Add files via upload
This is just a HW in GB. 
This code lets it find all jobs by keyword in HH.kz. Then it extracts job titles, links and salaries. This data is displayed on screen and is saved to the file named "hh_data.txt".  
Certainly, it has flaws including the ones I know. But I will be happy to hear (useful) comments from you.